Config.h 整理

+++ Software/X-Track/USER/App/Utils/TrackFilter/TrackPointFilter.cpp
@@ -284,7 +284,7 @@ double TrackPointFilter::QuickSort(double num)
     float x2, y;
     const float threehalfs = 1.5f;
     x2 = num * 0.5f;
-    y = num;
+    y = (float)num;
     i = *(long*)&y;
     i = 0x5f3759df - (i >> 1);
     y = *(float*)&i;
